pyproject.toml
[project]
name = "ai-news-aggregator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "apscheduler>=3.11.1",
    "beautifulsoup4>=4.14.2",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "openai>=2.8.1",
    "psycopg[binary]>=3.2.13",
    "pydantic>=2.12.4",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "sendgrid>=6.12.5",
    "sqlalchemy>=2.0.44",
    "tqdm>=4.67.1",
    "uvicorn>=0.38.0",
    "docling",
]
